ADD: Adds Regularity weights for each subject during atlas construction
The regularity weight will allow to give different weighting to different subjects
during the Atlas construction, such that the computed mean will be more closer (in
geodesic distance) to the subjects with higher weights. By default, the regularity
weights will be uniform such that the mean shape will be equidistant from all the
subjects. An addition field in deformable parameters XML file <regularity-weights>
needs to be provided with required weighting for each subject. If the number of provided
weights is not equal to the number of subjects, default uniform weighting is used.